Split event API into backend headers
Move the generic Event API to Kokkos_Event.hpp and place the CUDA specialization in Cuda/Kokkos_Cuda_Event.hpp. Extract the CUDA event handle as Impl::EventHandle<Cuda> and set the CUDA device around event operations.
