+++ core/src/Kokkos_Core.hpp
@@ -61,7 +61,7 @@
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
 
-#include <KokkosExp_Event.hpp>
+#include <Kokkos_Event.hpp>
 #include <Kokkos_Crs.hpp>
 #include <Kokkos_WorkGraphPolicy.hpp>
 // Including this in Kokkos_Parallel_Reduce.hpp led to a circular dependency
